Move pkg/bscatalog to pkg/output/bscatalog
