+++ pkg/output/bscatalog/v1alpha1/component.go
