+++ pkg/output/bscatalog/v1alpha1/constants.go
