+++ pkg/output/bscatalog/v1alpha1/group.go
