+++ pkg/output/catalog/component/component.go
@@ -6,7 +6,7 @@ import (
 	"fmt"
 	"time"
 
-	bscatalog "github.com/giantswarm/backstage-catalog-importer/pkg/bscatalog/v1alpha1"
+	bscatalog "github.com/giantswarm/backstage-catalog-importer/pkg/output/bscatalog/v1alpha1"
 )
 
 // Component holds our internal representation of something that we want
